# docs/api/qhse/openapi.yaml
openapi: 3.0.0
info:
  title: QHSE Module API
  description: Comprehensive API for Quality, Health, Safety, and Environment management
  version: 1.0.0
  contact:
    name: BlueDXP Platform
    email: support@bluedxp.com
servers:
  - url: /api/qhse
    description: QHSE API Base URL
...
paths:
...
  /search:
    get:
      tags: [Search]
      summary: Search QHSE data
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: types
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Search results
...